refactor out repeated usage of handleNeo4jInt
